hiroshima-news/requirements.txt
streamlit
openai
requests
beautifulsoup4
python-dotenv
psycopg2-binary
